Move Product Component to its folder
