+++ web/carousel/src/components/Product/ProductInSearch.module.css
